requirements.txt
gradio
torch
transformers
python-dotenv
pillow